requirements.txt
ray==2.7.0rc0
torch==2.0.0
torchvision==0.15.1
torchaudio==2.0.1
git+https://github.com/huggingface/transformers.git@df04959
deepspeed==0.10.0
fairscale==0.4.13
peft==0.5.0
datasets==2.14.4
accelerate==0.21.0
evaluate==0.4.0
bitsandbytes==0.41.1
wandb==0.15.8
pytorch-lightning==2.0.6
protobuf<3.21.0
torchmetrics==1.0.3
lm_eval==0.3.0
tiktoken==0.1.2
sentencepiece==0.1.99
urllib3<1.27
dm-tree